motorroutes/requirements.txt
amqp==5.0.6
billiard==3.6.4.0
cachetools==4.2.2
celery==5.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
Django==2.2.10
django-appconf==1.0.4
django-braces==1.14.0
django-extensions==3.1.3
django-filter==2.4.0
django-imagekit==4.0.2
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.1
google-auth==1.31.0
google-auth-oauthlib==0.4.4
idna==2.10
jsonfield==3.1.0
jwcrypto==0.9.1
kombu==5.1.0
oauthlib==3.1.1
pilkit==2.0
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
